# [Xamarin.Android.Build.Tasks] parse JDK `release` file directly