@jonathanpeppers 2024-01-22
> 3 hours and 46 minutes

I might have misread and this is 3 minutes 46 seconds